[CHJ]Update AllList_220222

+++ HSChat_Client/CFriendListForm.cpp
@@ -85,7 +85,7 @@ void CFriendListForm::OnBnClickedButtonFriendsDelete()
 			root["fnickname"] = std::string(CT2CA(fnickname));
 
 			m_pDlg->m_pClient->m_data.msg = writer.write(root);
-			m_pDlg->m_pClient->m_data.size = m_pDlg->m_pClient->m_data.msg.size();
+			m_pDlg->m_pClient->m_data.size = static_cast<int>(m_pDlg->m_pClient->m_data.msg.size());
 
 			m_pDlg->m_pClient->m_SendData();
