+++ HSChat_Client/CSearchPWForm.cpp
@@ -88,7 +88,7 @@ void CSearchPWForm::OnBnClickedButtonSearchPWOK()
 		root["id"] = std::string(CT2CA(strID));
 
 		m_pDlg->m_pClient->m_data.msg = writer.write(root);
-		m_pDlg->m_pClient->m_data.size = m_pDlg->m_pClient->m_data.msg.size();
+		m_pDlg->m_pClient->m_data.size = static_cast<int>(m_pDlg->m_pClient->m_data.msg.size());
 
 		m_pDlg->m_pClient->m_SendData();
